requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
livekit-api
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0

# Optional: Database support (uncomment when needed)
# asyncpg==0.29.0
# databases[postgresql]==0.8.0
# databases[sqlite]==0.8.0